refactor: split services into api and impl packages
