+++ src/main/java/com/uade/tpo/almacen/controller/CarritoController.java
@@ -5,8 +5,8 @@ import com.uade.tpo.almacen.entity.Usuario;
 import com.uade.tpo.almacen.entity.dto.CarritoResponse;
 import com.uade.tpo.almacen.excepciones.NoEncontradoException;
 import com.uade.tpo.almacen.mapper.CarritoMapper;
-import com.uade.tpo.almacen.service.CarritoService;
-import com.uade.tpo.almacen.service.UsuarioService;
+import com.uade.tpo.almacen.service.api.CarritoService;
+import com.uade.tpo.almacen.service.impl.UsuarioService;
 import jakarta.validation.constraints.Min;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
