+++ src/main/java/com/uade/tpo/almacen/controller/CategoriaController.java
@@ -10,7 +10,7 @@ import com.uade.tpo.almacen.entity.Categoria;
 import com.uade.tpo.almacen.controller.dto.CategoriaResponse; // ✅ DTO correcto
 import com.uade.tpo.almacen.excepciones.NoEncontradoException;
 import com.uade.tpo.almacen.excepciones.ParametroFueraDeRangoException;
-import com.uade.tpo.almacen.service.CategoriaService;
+import com.uade.tpo.almacen.service.api.CategoriaService;
 
 import java.net.URI;
 import java.util.List;
