+++ src/main/java/com/uade/tpo/almacen/service/api/CarritoService.java
@@ -1,0 +1,10 @@
+package com.uade.tpo.almacen.service.api;
+
+import com.uade.tpo.almacen.entity.Carrito;
+
+public interface CarritoService {
+    Carrito obtenerOCrearCarrito(int usuarioId);
+    Carrito agregarItem(int usuarioId, int productoId, int cantidad);
+    Carrito quitarItem(int usuarioId, int itemId);
+    void vaciarCarrito(int usuarioId);
+}
